Identify Key DataRefs for Pilot Performance #2
Documented and researched the key DataRefs of Altitude, Airspeed, Vertical Airspeed, and Heading. I have some concerns that I will address during our in-class code review, including the validatiy of the data I found and what the issue meant by find the usage of these DataRefs. These are listed in the notes in my markdown file.