updated Documentation for Sprint1
Updated the documentation and resolved comments I recieved from Prof. Holdener and my tech lead that there was redundency in the DataRefs listed (and that I should specifiy where and why these would be used), that I should document the instrument-based DataRefs of XPlane, that I should not submit questions with my commit request inside the documentation, and that I needed to put a Usage section in each DataRef information.