adding basic headings plugin
Copied from the 'get_altitudes.cpp' file that Ali Elnour originally wrote, this is a basic plugin initialization that deals with the Heading DataRefs for the Pilot and CoPilot's instruments. This has not been tested, as I am still trying to get a Windows VM on my Mac laptop, but nothing in the class has been changed except for the DataRef names and a function name (that of drawHelloWorld).