Initial Code Linting successful
The code linter is able to lint the changed files correctly, but does not check the entire project due to an ENV variable I have turned off. Next on my list of things to do is to have the code linter check the entire PilotDataSync project's C++, CLANG formatting, and Markdown files so they all 1) can run (though I may need to ditch the C++ correctness checker due to the XPlane stuff) and 2) look the same.