Relay Support for IPC reconnection
This commit deals with adding support on Relay's end for the Baton plugin reconnecting to the server. This is UNTESTED, however, as some of the implementation of Baton needs to be changed to support this reconnection mechanism (i.e. the changing of sending floats over IPC to strings). Thus, this code is untested, but does not seem to break the application.